third_party/WebKit/LayoutTests/accessibility/whitespace-in-name-calc-expected.txt

HelloWorld HelloWorld HelloWorld HelloWorld HelloWorld
HelloWorld
Hello
World
 
Hello
World 
Hello
World

Hello
World
Hello
World
Hello
World
PASS accessibleTitle("n1") is "HelloWorld"
PASS accessibleTitle("n2") is "HelloWorld"
PASS accessibleTitle("n3") is "HelloWorld"
PASS accessibleTitle("n4") is "HelloWorld"
PASS accessibleTitle("n5") is "HelloWorld"
PASS accessibleTitle("n6") is "HelloWorld"
PASS accessibleTitle("y1") is "Hello World"
PASS accessibleTitle("y2") is "Hello World"
PASS accessibleTitle("y3") is "Hello World"
PASS accessibleTitle("y4") is "Hello World"
PASS accessibleTitle("y5") is "Hello World"
PASS accessibleTitle("y6") is "Hello World"
PASS successfullyParsed is true

TEST COMPLETE
